[fix] #4080 - Move `base64` encode/decoder object to playground, remove from utils - Add test script and config to playground
