--- packages/playground/.babelrc
@@ -1,30 +1,0 @@
-{
-  "presets": [
-    [
-      "@babel/preset-env",
-      {
-        "modules": false, // don't convert modules into CommonJS
-        "useBuiltIns": "usage", // use polyfill when needed
-        "corejs": 3
-      }
-    ],
-    "@babel/preset-react"
-  ],
-  "plugins": [
-    "@babel/plugin-syntax-dynamic-import",
-    "@babel/plugin-proposal-object-rest-spread",
-    "@babel/plugin-proposal-class-properties",
-    [
-      "@babel/plugin-transform-runtime",
-      {
-        "corejs": 3,
-        "useESModules": true
-      }
-    ]
-  ],
-  "env": {
-    "development": {
-      "plugins": ["@babel/plugin-transform-react-jsx"]
-    }
-  }
-}
