+++ packages/playground/babel.config.js
@@ -1,0 +1,3 @@
+const defaultConfig = require('../../babel.config');
+
+module.exports = defaultConfig;
