+++ packages/playground/jest.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  verbose: true,
+  testEnvironment: 'jsdom',
+  testMatch: ['**/test/**/*.test.[jt]s?(x)'],
+  transformIgnorePatterns: [`/node_modules/(?!nanoid)`],
+};
