+++ packages/playground/package.json
@@ -3,7 +3,7 @@
   "version": "5.17.0",
   "description": "rjsf playground",
   "private": true,
-  "type": "module",
+  "type": "commonjs",
   "scripts": {
     "cs-check": "prettier -l \"src/**/*.ts?(x)\"",
     "cs-format": "prettier \"src/**/*.ts?(x)\" --write",
@@ -13,7 +13,8 @@
     "publish-to-gh-pages": "npm run build && gh-pages --dist build/",
     "publish-to-npm": "npm run build && npm publish",
     "start": "vite --force",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "test": "jest"
   },
   "lint-staged": {
     "src/**/*.ts?(x)": [
@@ -100,6 +101,8 @@
     "gh-pages": "^3.2.3",
     "html": "^1.0.0",
     "html-webpack-plugin": "^5.6.0",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "loader-utils": "^3.2.1",
     "mini-css-extract-plugin": "^2.8.0",
     "prettier": "^2.8.8",
