+++ packages/utils/src/index.ts
@@ -51,7 +51,6 @@ import utcToLocal from './utcToLocal';
 import validationDataMerge from './validationDataMerge';
 import withIdRefPrefix from './withIdRefPrefix';
 import getOptionMatchingSimpleDiscriminator from './getOptionMatchingSimpleDiscriminator';
-import base64 from './base64';
 
 export * from './types';
 export * from './enums';
@@ -121,5 +120,4 @@ export {
   utcToLocal,
   validationDataMerge,
   withIdRefPrefix,
-  base64,
 };
